updates for v1.0.7 - copy funcs into sep file & add output pix count csv
